refactor(lib): move all run_* subcommand handlers into commands.rs (#217)
Extract the 17 CLI subcommand handlers (run_sync, run_add, run_list, run_browse,
run_set, run_edit, run_generate, run_doctor, run_update, run_remove, run_tune,
run_log, run_profile, run_config, run_init, run_clean, run_completion) out of
the lib.rs monolith into a single src/commands.rs (~3.8k lines). The dispatcher
(run_cli / run), the clap Cli/Commands definitions, and all shared helpers stay
in lib.rs; commands reach them via crate:: glob imports.

lib.rs: 8222 -> 4408 lines. No visibility churn needed — commands is a
descendant module, so it already sees lib.rs's private helpers; only the
module-level use-imports were rebalanced between the two files.

Rebased on top of #229 (incremental generate); the new run_generate(force)
perf code moves across unchanged. The lib.rs mod tests block still references
url/merge/plugin_build helpers by bare name, so those globs are imported in the
test module (relocation tracked in #224).

Behaviour-preserving code motion; CLI --help / completion smoke-checked. fmt
--check + clippy -D warnings + test (823) + --doc all green.
